% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTShiny.R
\name{CTShiny}
\alias{CTShiny}
\title{Launch 'CTShiny' Interface}
\usage{
CTShiny()
}
\value{
Nothing
}
\description{
CTShiny() loads interactive user interface built using R 'shiny'.
}
\details{
Start CTShiny

The interactive user interface is to provide an easy way for classification tree analysis and downloading relevant plot.
}
\examples{
if(interactive()){
library(rmarkdown)
CTShiny()
}
}
\keyword{CTShiny}
